// my-app/src/types/store/actionTypes.ts
export enum actionTypes {
  LOADING_START = 'LOADING_START',
  LOADING_STOP = 'LOADING_STOP',
  INPUT = 'INPUT',
  FORM = 'FORM',
  SORT = 'SORT',
  PER_PAGE = 'PER_PAGE',
  FETCH_CARDS = 'FETCH_CARDS',
  FETCH_EMPTY = 'FETCH_EMPTY',
  CHANGE_PAGE = 'CHANGE_PAGE',
  PICK_CARD = 'PICK_CARD',
  UNPICK_CARD = 'UNPICK_CARD',
}